+++ src/frontend/js/hooks/useTeacherPendingContractsCount/index.ts
@@ -1,6 +1,4 @@
-import { useMemo } from 'react';
 import { useOrganizationContracts } from 'hooks/useContracts';
-import { useCourseProductRelation } from 'hooks/useCourseProductRelation';
 import { PER_PAGE } from 'settings';
 import { ContractState, CourseProductRelation, Organization } from 'types/Joanie';
 
@@ -13,30 +11,15 @@ const useTeacherPendingContractsCount = ({
   organizationId,
   courseProductRelationId,
 }: UseTeacherPendingContractsCountProps) => {
-  const {
-    item: training,
-    states: { isFetched: isTrainingFetched },
-  } = useCourseProductRelation(courseProductRelationId, {
+  const { items: contracts, meta } = useOrganizationContracts({
     organization_id: organizationId,
+    course_product_relation_id: courseProductRelationId,
+    signature_state: ContractState.LEARNER_SIGNED,
+    page: 1,
+    page_size: PER_PAGE.teacherContractList,
   });
 
-  const isActive = useMemo(() => {
-    return !!(organizationId && (!courseProductRelationId || isTrainingFetched));
-  }, [organizationId, courseProductRelationId, isTrainingFetched]);
-
-  const { items: contracts, meta } = useOrganizationContracts(
-    {
-      organization_id: organizationId,
-      course_id: training?.course.id,
-      product_id: training?.product.id,
-      signature_state: ContractState.LEARNER_SIGNED,
-      page: 1,
-      page_size: PER_PAGE.teacherContractList,
-    },
-    { enabled: isActive },
-  );
-
-  if (isActive) {
+  if (organizationId) {
     return {
       contracts,
       pendingContractCount: meta?.pagination?.count ?? 0,