+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Contract list in the teacher dashbaord are now filtered by 
+  courseProductRelationId instead of courseId and productId.
 - Switch from setup.cfg to pyproject.toml 
 
 ## [2.25.0-beta.1]